divide now does decimal points

+++ src/main/java/org/example/Calculator/Calculator.java
@@ -15,7 +15,7 @@ public class Calculator
         return i - i1;
     }
 
-    public int divide(int i, int i1) {
-        return i / i1;
+    public double divide(double i, double i1) {
+            return i / i1;
     }
 }
